# api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateUserInputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "/api/schemas/CreateUserInputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "login": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "login",
          "name",
          "password"
        ],
        "type": "object"
      },
      "CreatedOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "/api/schemas/CreatedOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "id"
        ],
        "type": "object"
      },
...
      "EmailInputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "/api/schemas/EmailInputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "email": {
            "description": "Null or blank removes it",
            "maxLength": 254,
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "/api/schemas/Error.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "params": {
            "additionalProperties": {},
            "type": "object"
          }
        },
        "required": [
          "code"
        ],
        "type": "object"
      },
...
      "Grant": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "org_unit_id": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "org_unit_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "product": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "product",
          "role",
          "org_unit_id",
          "org_unit_name"
        ],
        "type": "object"
      },
      "GrantInputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "/api/schemas/GrantInputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "org_unit_id": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "product": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "required": [
          "product",
          "role",
          "org_unit_id"
        ],
        "type": "object"
      },
...
      "LoginInputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "/api/schemas/LoginInputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "login": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "login",
          "password"
        ],
        "type": "object"
      },
...
      "Me": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "/api/schemas/Me.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "authz_version": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "locale": {
            "enum": [
              "vi",
              "en"
            ],
            "type": "string"
          },
          "login": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": [
                "array",
                "null"
              ]
            },
            "type": "object"
          },
          "products": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "products_with_data": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "timezone": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "login",
          "name",
          "locale",
          "timezone",
          "authz_version",
          "products",
          "products_with_data",
          "permissions"
        ],
        "type": "object"
      },
...
      "OrgUnit": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "maxLength": 500,
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "kind": {
            "enum": [
              "group",
              "company",
              "branch",
              "department"
            ],
            "type": "string"
          },
          "legal_name": {
            "maxLength": 300,
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "parent_id": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "tax_code": {
            "maxLength": 20,
            "type": "string"
          }
        },
        "required": [
          "id",
          "parent_id",
          "kind",
          "name"
        ],
        "type": "object"
      },
      "OrgUnitInput": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "/api/schemas/OrgUnitInput.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "address": {
            "maxLength": 500,
            "type": "string"
          },
          "kind": {
            "enum": [
              "group",
              "company",
              "branch",
              "department"
            ],
            "type": "string"
          },
          "legal_name": {
            "maxLength": 300,
            "type": "string"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "parent_id": {
            "format": "int64",
            "type": [
              "integer",
              "null"
            ]
          },
          "tax_code": {
            "maxLength": 20,
            "type": "string"
          }
        },
        "required": [
          "parent_id",
          "kind",
          "name"
        ],
        "type": "object"
      },
...
      "Role": {
        "additionalProperties": false,
        "properties": {
          "permissions": {
            "items": {
              "type": "string"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "product": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "tenant_wide": {
            "type": "boolean"
          }
        },
        "required": [
          "product",
          "role",
          "permissions",
          "tenant_wide"
        ],
        "type": "object"
      },
...
      "UpdateMeInputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "/api/schemas/UpdateMeInputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "locale": {
            "enum": [
              "vi",
              "en"
            ],
            "type": "string"
          }
        },
        "required": [
          "locale"
        ],
        "type": "object"
      },
      "User": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "/api/schemas/User.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "email": {
            "description": "Where email notifications go; null for none",
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "login": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "login",
          "name",
          "email"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "mmerp",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "parameters": [
          {
            "in": "cookie",
            "name": "session",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      }
    },
...
    "/me": {
      "get": {
        "operationId": "get-me",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      },
      "patch": {
        "operationId": "update-me",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      }
    },
...
    "/org-units": {
      "get": {
        "operationId": "list-org-units",
        "parameters": [
          {
            "description": "With permission: only units where the actor has it",
            "explode": false,
            "in": "query",
            "name": "product",
            "schema": {
              "description": "With permission: only units where the actor has it",
              "type": "string"
            }
          },
          {
            "explode": false,
            "in": "query",
            "name": "permission",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrgUnit"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      },
      "post": {
        "operationId": "create-org-unit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUnitInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedOutputBody"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      }
    },
    "/org-units/{id}": {
      "put": {
        "operationId": "update-org-unit",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUnitInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      }
    },
...
    "/roles": {
      "get": {
        "operationId": "list-roles",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "list-users",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      },
      "post": {
        "operationId": "create-user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedOutputBody"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "get-user",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      }
    },
    "/users/{id}/email": {
      "put": {
        "operationId": "set-user-email",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      }
    },
    "/users/{id}/roles": {
      "get": {
        "operationId": "list-user-roles",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Grant"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      },
      "post": {
        "operationId": "grant-role",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedOutputBody"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      }
    },
    "/users/{id}/roles/{grant}": {
      "delete": {
        "operationId": "revoke-role",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "grant",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error"
          }
        },
        "tags": [
          "core",
          "core"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}